neighbors in 3D

+++ Neighbor.h
@@ -19,7 +19,7 @@ public:
     
     Neighbor(void){};
     void place_particles(vector<Particle*>&,float,Container);
-    int compute_box_num(Vec3,float,int,int);
+    int compute_box_num(Vec3,float,float,float);
     void add_to_box_particles(int,int);
     void set_particle_neighbors(int, Particle*);
 };
